test: add loading plugins test
